refactor: ShoppingActivity 코드 모듈화
