+++ app/src/main/java/woowacourse/shopping/domain/Cart.kt
@@ -1,7 +1,5 @@
 package woowacourse.shopping.domain
 
-import woowacourse.shopping.presentation.ui.Product
-
 data class Cart(
     val product: Product,
     val count: Int = DEFAULT_PURCHASE_COUNT,
