+++ app/src/main/java/woowacourse/shopping/domain/Product.kt
@@ -1,0 +1,8 @@
+package woowacourse.shopping.domain
+
+data class Product(
+    val id: Long,
+    val name: String,
+    val imgUrl: String,
+    val price: Long,
+)
