+++ app/src/main/java/woowacourse/shopping/domain/ProductRepository.kt
@@ -1,7 +1,5 @@
 package woowacourse.shopping.domain
 
-import woowacourse.shopping.presentation.ui.Product
-
 interface ProductRepository {
     fun load(
         pageOffset: Int,
