+++ app/src/main/java/woowacourse/shopping/presentation/ui/cart/CartHandler.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.presentation.ui.cart
 
-import woowacourse.shopping.presentation.ui.Product
+import woowacourse.shopping.domain.Product
 
 interface CartHandler {
     fun onDeleteClick(product: Product)
