+++ app/src/main/java/woowacourse/shopping/presentation/ui/cart/CartViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import woowacourse.shopping.domain.Cart
 import woowacourse.shopping.domain.CartRepository
-import woowacourse.shopping.presentation.ui.Product
+import woowacourse.shopping.domain.Product
 import woowacourse.shopping.presentation.ui.UiState
 
 class CartViewModel(private val cartRepository: CartRepository) : ViewModel() {
