+++ app/src/main/java/woowacourse/shopping/presentation/ui/detail/ProductDetailActivity.kt
@@ -20,6 +20,7 @@ class ProductDetailActivity : BindingActivity<ActivityProductDetailBinding>() {
     private val viewModel: ProductDetailViewModel by viewModels { ViewModelFactory() }
 
     override fun initStartView() {
+        setTitle("Detail")
         val id = intent.getLongExtra(EXTRA_PRODUCT_ID, -1L)
         if (id == -1L) finish()
         viewModel.loadById(id)
