+++ app/src/main/java/woowacourse/shopping/presentation/ui/shopping/adapter/ShoppingViewHolder.kt
@@ -1,0 +1,46 @@
+package woowacourse.shopping.presentation.ui.shopping.adapter
+
+import android.view.View
+import androidx.recyclerview.widget.RecyclerView
+import com.bumptech.glide.Glide
+import woowacourse.shopping.databinding.ItemLoadBinding
+import woowacourse.shopping.databinding.ItemProductBinding
+import woowacourse.shopping.domain.Product
+import woowacourse.shopping.presentation.ui.getPriceText
+import woowacourse.shopping.presentation.ui.shopping.ShoppingHandler
+
+sealed class ShoppingViewHolder(view: View) : RecyclerView.ViewHolder(view) {
+    class ProductViewHolder(
+        private val binding: ItemProductBinding,
+        private val shoppingHandler: ShoppingHandler,
+    ) :
+        ShoppingViewHolder(binding.root) {
+        private var id: Long = -1
+
+        init {
+            binding.root.setOnClickListener {
+                shoppingHandler.onClick(id)
+            }
+        }
+
+        fun bind(item: Product) {
+            id = item.id
+            binding.tvName.text = item.name
+            binding.tvPrice.text = getPriceText(binding.root.context, item.price)
+            Glide.with(binding.root.context)
+                .load(item.imgUrl)
+                .into(binding.imgItem)
+        }
+    }
+
+    class LoadViewHolder(
+        private val binding: ItemLoadBinding,
+        private val shoppingHandler: ShoppingHandler,
+    ) : ShoppingViewHolder(binding.root) {
+        fun bind() {
+            binding.btnShowMore.setOnClickListener {
+                shoppingHandler.loadMore()
+            }
+        }
+    }
+}
