+++ app/src/main/res/layout/activity_shopping.xml
@@ -2,14 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools">
-
-    <data>
-
-        <variable
-            name="vm"
-            type="woowacourse.shopping.presentation.ui.shopping.ShoppingViewModel" />
-
-    </data>
+    <data/>
 
     <androidx.constraintlayout.widget.ConstraintLayout
